backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
google-generativeai